requirements.txt
python-dotenv
graphrag_sdk[all]
kubemq>=3.4.1